Update patch set 2
Patch Set 2:

CFX appears to use AOSP clang/llvm from AOSP with Linaro cherry-picks from SM.  The Krait commits here are cherry-picked straight from the LLVM mirror at github

https://github.com/llvm-mirror/llvm/commit/a2bfea97432fe5edc153f1556a6e1c8fb0e5e67c

https://github.com/llvm-mirror/clang/commit/69a9e166e083131c91409207a5ee07479c1cc14d

so I think it'd be awesome to build with that maybe with AOSP/Linaro patches on top.  Although, it is experimental, so not sure how stable it is and how suited it'd be for Omni.  

It looks like LLVM/Clang are the future https://collaborate.linaro.org/pages/viewrecentblogposts.action?key=LLVM

AOSP is horrendous at staying in line with upstream sources (CAF, tinyalsa, llvm/clang, etc) so it'd be nice to have the most up-to-date clang/llvm when it replaces GCC.

I'd be happy to work with you.  Add me on Hangouts (xboxlover360@gmail.com)

Patch-set: 2